Clean out all the warnings. TODO: Some of the changes, especially in the Serializer code, might have run time implications. Revisit this during testing TODO

+++ antlrcpp/BitSet.h
@@ -43,7 +43,7 @@ public:
 
 	int nextSetBit(const int & pos) {
 		for (size_t i = pos; i < data.size(); i++){
-			if (data.test(i)) return i;
+			if (data.test(i)) return (int)i;
 		}
 
 		return -1;
