+++ antlrcpp/CPPUtils.h
@@ -56,7 +56,7 @@ namespace antlrcpp {
     
     std::wstring stringFormat(const wchar_t* fmt, ...);
 
-	wchar_t* toCharArray(const std::vector<int> *data);
+	wchar_t* toCharArray(const std::vector<size_t> *data);
 
 	std::wstring toHexString(const int t);
